requirements.txt
chainlit==1.3.0rc1
openai
yfinance
plotly
websockets
psycopg2-binary
# Utilities
python-dotenv
tqdm
gunicorn
elasticsearch
llama-index
llama-index-vector-stores-elasticsearch
llama-index-embeddings-openai
openai
pandas
numpy
tqdm